+++ packages/i18n-typed-store/src/index.ts
@@ -2,7 +2,7 @@
 export { createTranslationModuleMap } from './lib/create-translation-module-map';
 export { createTranslationStore } from './lib/create-translation-store';
 export { createPluralSelector } from './lib/create-plural-selector';
-export { getTranslation } from './lib/get-translation';
+export { getTranslation, getTranslationOrThrow, TranslationMissingError } from './lib/get-translation';
 export { parseLocale, generateLocaleCandidates, findBestLocaleMatch } from './lib/locale-utils';
 export { interpolate } from './lib/interpolate';
 export { createIntlFormatters } from './lib/intl-formatters';